Migrate fjs/sul/id to .f.mjs, split types into types.ts
Splits the Id type into a sibling types.ts, and converts the runtime
implementation to JSDoc-typed module.f.mjs/proof.f.mjs under the
stage-1 TypeScript-to-mjs migration. Updates the four dependents
(fjs/sul/module.f.ts, fjs/sul/proof.f.ts, fjs/sul/level/hash/module.f.ts,
fjs/sul/level/hash/proof.f.ts) to import runtime values from
module.f.mjs and the Id type from types.ts.
