+++ fjs/sul/id/module.f.mjs
@@ -2,29 +2,24 @@
  * 256-bit content-addressed identifiers for SUL values.
  * Every identifier is one of three variants: a level-3 literal (inline), a raw bit vector (inline),
  * or a SHA2-based hash, selected automatically by `compress` based on size and input type.
+ * See `./types.ts` for the `Id` type.
  *
  * @module
  */
 
 import { toArray } from '../../types/list/module.f.mjs'
-import type { Vec } from '../../types/bit_vec/types.ts'
+/** @import { Vec } from '../../types/bit_vec/types.ts' */
 import { length, msb, uint, uintChunkList, unpack, vec } from '../../types/bit_vec/module.f.mjs'
 import { assertEq } from '../../asserts/module.f.mjs'
 import { utf8 } from '../../text/module.f.mjs'
-import type { Point2D } from '../../crypto/secp/types.ts'
+/** @import { Point2D } from '../../crypto/secp/types.ts' */
 import { secp256r1 } from '../../crypto/secp/module.f.mjs'
-import type { V8 } from '../../crypto/sha2/types.ts'
+/** @import { V8 } from '../../crypto/sha2/types.ts' */
 import { base32 } from '../../crypto/sha2/module.f.mjs'
 import { literal3ToVec } from '../level/literal/module.f.mjs'
 import { log2 } from '../../types/bigint/module.f.mjs'
-import type { Nominal } from '../../types/nominal/types.ts'
 import { asBase, asNominal } from '../../types/nominal/module.f.mjs'
-
-/** A 256-bit SUL identifier. One of three variants: level-3 literal, raw bit vector, or SHA2 hash. */
-export type Id = Nominal<
-    'sul/id',
-    '6f5f6da053a6ac70e9687d42b7a09e925c3be21027f55beb2cba3040bf3d5b71',
-    bigint>
+/** @import { Id } from './types.ts' */
 
 // 32 bytes = 256 bits.
 //
@@ -36,7 +31,8 @@ const utf8IvSeed = utf8(ivSeed)
 
 const c = secp256r1
 
-const ivUint: bigint = (c.mul(uint(utf8IvSeed))(c.g) as Point2D)[0]
+/** @type {bigint} */
+const ivUint = /** @type {Point2D} */ (c.mul(uint(utf8IvSeed))(c.g))[0]
 
 // 64 hex = 256 bits = 32 bytes:
 assertEq(
@@ -46,14 +42,16 @@ assertEq(
     0x325d5666_573eb118_f32191de_20d17f64_33392ba3_291ae46c_1474a5ed_a5383f25n
 )
 
-const iv = toArray(uintChunkList(msb)(32n)({ length: 256n, uint: ivUint })) as V8
+const iv = /** @type {V8} */ (toArray(uintChunkList(msb)(32n)({ length: 256n, uint: ivUint })))
 
 assertEq(iv.length, 8)
 
 /**
  * Note: no need to add a prefix.
+ *
+ * @type {(v: bigint) => Id}
  */
-export const level3Id: (v: bigint) => Id =
+export const level3Id =
     asNominal
 
 const rawPrefixOffset = 0xFEn
@@ -67,24 +65,26 @@ assertEq(
     0x40000000_00000000_00000000_00000000_00000000_00000000_00000000_00000000n
 )
 
+// 253
+const rawLenMax = 0xFDn
+
 /**
  * Note: length(symbol) <= 253n
  *
- * @param symbol
- * @returns
+ * @param {Vec} symbol
+ * @returns {Id}
  */
-export const rawId = (symbol: Vec): Id => {
+export const rawId = symbol => {
     const { length, uint } = unpack(symbol)
     return asNominal(rawPrefix | uint | (1n << length))
 }
 
-// 253
-const rawLenMax = 0xFDn
-
-export const isRaw = (v: Id): boolean =>
+/** @type {(v: Id) => boolean} */
+export const isRaw = v =>
     asBase(v) >> rawPrefixOffset === 1n
 
-const toRaw = (a: Id): Vec => {
+/** @type {(a: Id) => Vec} */
+const toRaw = a => {
     if (!isRaw(a)) {
         return literal3ToVec(asBase(a))
     }
@@ -104,17 +104,18 @@ assertEq(
     0x80000000_00000000_00000000_00000000_00000000_00000000_00000000_00000000n
 )
 
-export const isHash = (v: Id): boolean =>
+/** @type {(v: Id) => boolean} */
+export const isHash = v =>
     asBase(v) >> hashPrefixOffset === 1n
 
 /**
  * Note: we don't need to remove the prefix bits from the hash because
  * the prefix equals the prefix mask (`1n << 0xFFn`).
  *
- * @param symbol
- * @returns
+ * @param {bigint} hash
+ * @returns {Id}
  */
-export const hashId = (hash: bigint): Id =>
+export const hashId = hash =>
     asNominal(hashPrefix | hash)
 
 const hash2 = base32.compress(iv)
@@ -123,10 +124,12 @@ const vecX20 = vec(0x20n)
 
 const { concat, listToVec } = msb
 
-const hashMerge = (a: Id, b: Id): Id =>
+/** @type {(a: Id, b: Id) => Id} */
+const hashMerge = (a, b) =>
     hashId(uint(listToVec(hash2((asBase(a) << 0x100n) | asBase(b)).map(vecX20))))
 
-export const compress = (a: Id, b: Id): Id => {
+/** @type {(a: Id, b: Id) => Id} */
+export const compress = (a, b) => {
     if (isHash(a) || isHash(b)) {
         return hashMerge(a, b)
     }
