+++ fjs/sul/proof.f.ts
@@ -1,5 +1,6 @@
 import { assert, assertEq } from '../asserts/module.f.mjs'
-import { compress, type Id } from './id/module.f.ts'
+import { compress } from './id/module.f.mjs'
+import type { Id } from './id/types.ts'
 import { encode, emptyEncodeState } from './module.f.ts'
 import type { Add } from './level/hash/module.f.ts'
